assig1/grpc_client/requirements.txt
grpcio
googleapis-common-protos
python-benedict
requests
